Move quaternion related header files associated with app_potts_quaterion to a subdir 'potts_quaternion'
