+++ src/potts_quaternion/cubic_symmetries.h
