+++ src/potts_quaternion/disorientation.h
