+++ src/potts_quaternion/hcp_symmetries.h
