+++ src/potts_quaternion/quaternion.h
